User creation, validation and authentication.

+++ server/database/migrations/1617422476424_messages.ts
@@ -18,6 +18,6 @@ export default class Messages extends BaseSchema {
     }
   }
   public async down() {
-    this.schema.dropTable(this.tableName)
+    if (await this.schema.hasTable(this.tableName)) this.schema.dropTable(this.tableName)
   }
 }
