+++ server/database/migrations/1617423755677_certifications.ts
@@ -18,6 +18,6 @@ export default class Certifications extends BaseSchema {
   }
 
   public async down() {
-    this.schema.dropTable(this.tableName)
+    if (await this.schema.hasTable(this.tableName)) this.schema.dropTable(this.tableName)
   }
 }
