+++ server/database/migrations/1617424782741_level_of_educations.ts
@@ -15,6 +15,6 @@ export default class LevelOfEducations extends BaseSchema {
     }
   }
   public async down() {
-    this.schema.dropTable(this.tableName)
+    if (await this.schema.hasTable(this.tableName)) this.schema.dropTable(this.tableName)
   }
 }
